Refactoring

+++ internal/database/village_setup.go
@@ -11,7 +11,7 @@ type VillageSetupModel struct {
 	DB *sqlx.DB
 }
 
-func (m *VillageSetupModel) Insert(newVillageSetup models.VillageSetup) (int, error) {
+func (m *VillageSetupModel) Insert(newVillageSetup models.VillageSetup) (uint32, error) {
 
 	stmt := `INSERT OR IGNORE INTO village_setup (village_id, player_id, hunterhut_1, hunterhut_2, hunterhut_3, hunterhut_4, hunterhut_5, woodcutterhut_1, woodcutterhut_2, woodcutterhut_3, woodcutterhut_4, woodcutterhut_5,quarry_1, quarry_2, quarry_3, quarry_4, quarry_5, coppermine_1, coppermine_2, coppermine_3, coppermine_4, coppermine_5, fountain_1, fountain_2, fountain_3, fountain_4, fountain_5) VALUES (:village_id, :player_id, :hunterhut_1, :hunterhut_2, :hunterhut_3, :hunterhut_4, :hunterhut_5, :woodcutterhut_1, :woodcutterhut_2, :woodcutterhut_3, :woodcutterhut_4, :woodcutterhut_5, :quarry_1, :quarry_2, :quarry_3, :quarry_4, :quarry_5, :coppermine_1, :coppermine_2, :coppermine_3, :coppermine_4, :coppermine_5, :fountain_1, :fountain_2, :fountain_3, :fountain_4, :fountain_5);`
 
@@ -26,13 +26,13 @@ func (m *VillageSetupModel) Insert(newVillageSetup models.VillageSetup) (int, er
 		return 0, err
 	}
 
-	return int(id), nil
+	return uint32(id), nil
 }
 
-func (m *VillageSetupModel) InsertWithIDCheck(village_id int, player_id int32) (int, error) {
+func (m *VillageSetupModel) InsertWithIDCheck(village_id uint32, player_id uint32) (uint32, error) {
 
 	newVillageSetup := models.VillageSetup{
-		VillageID:       int32(village_id),
+		VillageID:       village_id,
 		PlayerID:        player_id,
 		HunterHut_1:     1,
 		WoodcutterHut_1: 1,
@@ -54,10 +54,10 @@ func (m *VillageSetupModel) InsertWithIDCheck(village_id int, player_id int32) (
 		return 0, err
 	}
 
-	return int(id), nil
+	return uint32(id), nil
 }
 
-func (m *VillageSetupModel) Update(newVillageSetup models.VillageSetup) (int, error) {
+func (m *VillageSetupModel) Update(newVillageSetup models.VillageSetup) (uint32, error) {
 
 	stmt := `INSERT OR IGNORE INTO village_setup (village_id, player_id, hunterhut_1, hunterhut_2, hunterhut_3, hunterhut_4, hunterhut_5, woodcutterhut_1, woodcutterhut_2, woodcutterhut_3, woodcutterhut_4, woodcutterhut_5,quarry_1, quarry_2, quarry_3, quarry_4, quarry_5, coppermine_1, coppermine_2, coppermine_3, coppermine_4, coppermine_5, fountain_1, fountain_2, fountain_3, fountain_4, fountain_5) VALUES (:village_id, :player_id, :hunterhut_1, :hunterhut_2, :hunterhut_3, :hunterhut_4, :hunterhut_5, :woodcutterhut_1, :woodcutterhut_2, :woodcutterhut_3, :woodcutterhut_4, :woodcutterhut_5, :quarry_1, :quarry_2, :quarry_3, :quarry_4, :quarry_5, :coppermine_1, :coppermine_2, :coppermine_3, :coppermine_4, :coppermine_5, :fountain_1, :fountain_2, :fountain_3, :fountain_4, :fountain_5);`
 
@@ -72,5 +72,5 @@ func (m *VillageSetupModel) Update(newVillageSetup models.VillageSetup) (int, er
 		return 0, err
 	}
 
-	return int(id), nil
+	return uint32(id), nil
 }
